+++ examples/AllFunctions/AllFunctions.ino
@@ -261,7 +261,9 @@ void loop() {
 
 #if TINY_GSM_TEST_GPRS
   // Unlock your SIM card with a PIN if needed
-  if (GSM_PIN && modem.getSimStatus() != 3) { modem.simUnlock(GSM_PIN); }
+  if (GSM_PIN && static_cast<int>(modem.getSimStatus()) != 3) {
+    modem.simUnlock(GSM_PIN);
+  }
 #endif
 
 #if TINY_GSM_TEST_WIFI && defined(TINY_GSM_MODEM_HAS_WIFI)
